Add a single entrypoint with subcommands for all parts of almond-cloud
"almond-cloud foo" instead of finding wherever "foo.js" is defined.
Also supports "almond-cloud --help" and "almond-cloud foo --help",
which makes it easy to understand what each command does.

The single entrypoint also performs shared initialization, ensuring
every part of Almond has a consistent view of the configuration
and a consistent promise setup.

The name almond-cloud is chosen because "almond" is taken (by
almond-cmdline, and by <npmjs.org/almond>, an unrelated project)
