+++ almond-cloud@.service
@@ -6,7 +6,7 @@ User=thingengine
 WorkingDirectory=/srv/almond-cloud
 EnvironmentFile=/etc/sysconfig/almond-cloud
 ExecStartPre=/usr/bin/rm -f /srv/almond-cloud/control /srv/almond-cloud/direct
-ExecStart=/usr/bin/node /opt/almond-cloud/almond/master.js --shard %I
+ExecStart=/opt/almond-cloud/main.js run-almond --shard %I
 SyslogIdentifier=almond-cloud-master-%I
 CapabilityBoundingSet=CAP_SYS_ADMIN
 KillMode=mixed
