+++ package.json
@@ -15,6 +15,9 @@
       "email": "silei@cs.stanford.edu"
     }
   ],
+  "bin": {
+    "almond-cloud": "./main.js"
+  },
   "scripts": {
     "update-pot": "xgettext -f po/POTFILES -o po/${npm_package_name}.pot --from-code UTF-8 --package-name ${npm_package_name} --package-version ${npm_package_version}; xargs -a po/POTFILES.jade jsxgettext -k _,N_,gettext,ngettext,pgettext -L pug -o po/${npm_package_name}.pot -j",
     "update-po": "./build/update-po.sh",
