+++ production/training/k8/gpu-deployment.yaml
@@ -13,16 +13,13 @@ spec:
       - name: "gpu-training"
         image: ${REPO}/almond/training:v0.7
         imagePullPolicy: Always
-        env:
-        - name: "PORT"
-          value: "9999"
-        - name: "RUN_GPU_DAEMON"
-          value: "true"
+        command:
+          - /bin/bash
+          - /var/lib/almond-training/startup.sh
+          - run-gpu-training
         resources:
           limits:
               nvidia.com/gpu: 1
-        ports:
-        - containerPort: 9999
         volumeMounts:
         - name: config
           mountPath: "/etc/almond-cloud"
