+++ tests/nlp-integration.sh
@@ -8,6 +8,9 @@ set -o pipefail
 
 srcdir=`dirname $0`/..
 srcdir=`realpath $srcdir`
+almond-cloud() {
+	${srcdir}/main.js "$@"
+}
 
 DATABASE_URL="mysql://thingengine:thingengine@localhost/thingengine_test"
 export DATABASE_URL
@@ -64,8 +67,8 @@ mkdir -p $workdir/shared/cache
 echo '{"tt:stock_id:goog": "fb80c6ac2685d4401806795765550abdce2aa906.png"}' > $workdir/shared/cache/index.json
 
 # clean the database and bootstrap
-$srcdir/scripts/execute-sql-file.js $srcdir/model/schema.sql
-node $srcdir/scripts/bootstrap.js
+almond-cloud execute-sql-file $srcdir/model/schema.sql
+almond-cloud bootstrap
 
 mkdir -p 'org.thingpedia.models.default:en'
 mkdir -p 'org.thingpedia.models.contextual:en'
@@ -78,7 +81,7 @@ tar xvf $srcdir/tests/embeddings/current-contextual.tar.gz -C 'org.thingpedia.mo
 
 # remove developer models that were autoadded by bootstrap
 # we'll test the main models only (there is no difference really)
-$srcdir/scripts/execute-sql-file.js <(echo "delete from models where tag like '%developer%'")
+almond-cloud execute-sql-file <(echo "delete from models where tag like '%developer%'")
 
 mkdir -p 'classifier'
 wget --no-verbose -c https://nnmaster.almond.stanford.edu/test-models/classifier1.tar.gz -O $srcdir/tests/embeddings/classifier1.tar.gz
