+++ tests/thingpedia-integration.sh
@@ -9,6 +9,9 @@ set -o pipefail
 
 srcdir=`dirname $0`/..
 srcdir=`realpath $srcdir`
+almond-cloud() {
+	${srcdir}/main.js "$@"
+}
 
 PORT=${PORT:-8080}
 cat > $srcdir/secret_config.js <<EOF
@@ -58,14 +61,14 @@ echo '{"tt:stock_id:goog": "fb80c6ac2685d4401806795765550abdce2aa906.png"}' > $w
 # clean the database and bootstrap
 # (this has to occur after setting up the download
 # directories because it copies the icon png files)
-$srcdir/scripts/execute-sql-file.js $srcdir/model/schema.sql
-node $srcdir/scripts/bootstrap.js
+almond-cloud execute-sql-file $srcdir/model/schema.sql
+almond-cloud bootstrap
 
 # load some more data into Thingpedia
 test -f $srcdir/tests/data/com.bing.zip || wget https://thingpedia.stanford.edu/thingpedia/download/devices/com.bing.zip -O $srcdir/tests/data/com.bing.zip
 eval $(node $srcdir/tests/load_test_thingpedia.js)
 
-node $srcdir/frontend.js &
+almond-cloud run-frontend &
 frontendpid=$!
 
 # in interactive mode, sleep forever
@@ -116,10 +119,10 @@ EOF
 # too, so make sure we don't die with 400 or 500 because Almond is off
 # we have just tested operation without web almond anyway
 export THINGENGINE_DISABLE_SYSTEMD=1
-node $srcdir/almond/master.js &
+almond-cloud run-almond &
 masterpid=$!
 
-node $srcdir/frontend.js &
+almond-cloud run-frontend &
 frontendpid=$!
 
 if test "$1" = "--webalmond-interactive" ; then
